+++ roots.h
@@ -10,7 +10,7 @@ extern int  solve_quartic1(Flt *, Flt *, Flt, Flt);
 extern int  bounded_polysolve(int, Flt *, Flt *, Flt, Flt);
 extern int  Inside_Polygon(Flt, Flt, int, fVec *, int, int);
 extern int  Inside_Contour(Flt x, Flt y, int itype, int n, fVec *points);
-extern long binomial(int, int);
+extern int binomial(int, int);
 extern int  binomials[15][15];
 
 #endif /* __ROOT_SOLVER_DEFS */
